Catering for the 8MB case in psyqo's chained DMA.
The previous mechanism was a nice trick to allow for only a lui to load the constant required for the end marker, but it turns out this is a valid address in the 8MB range, so let's actually use the proper marker here instead. It shouldn't really affect codegen much.